==> Machine_Learning/Estudio Parque Automotriz/01 - ETL Parque.ipynb <==
## Base Parque Automotriz

# Import librerías

import pandas as pd

# Carga Base Parque

URL = 'C:/Users/juan.vergara/OneDrive - Caren Repuestos Flotacentro/Documentos/Análisis/Parque Automotriz/Parque en Bruto para Modelo/' 
URL_TRAD = 'C:/Users/juan.vergara/OneDrive - Caren Repuestos Flotacentro/Documentos/Análisis/Parque Automotriz/Parque en Bruto para Modelo/Traductores/' 

df_parque = pd.read_table(URL + 'BaseParque2020.txt', sep = '|')

df_parque.head(5)

df_parque['ppu'].count()

Campos: <br/>

**ppu**: Patente (Placa Patente Única) <br />
**dv_ppu**:Dígito verificador de la patente  <br />
**tipo**: Tipo del vehículo  <br />
**ano_fab**: Año de fabricación del vehículo <br />
**marca**: Marca del vehículo  <br />
**modelo**: Modelo del vehículo  <br />
**color**: Color del vehículo  <br />
**motor**: Motor del vehículo  <br />
**chasis**: Chasís del motor  <br />
**transfer**:  <br />
**rut_ant**:  <br />
**dv_ant**:  <br />
**rut**: Rut sin digito verificador del dueño  <br />
**dv**: Dígito verificador del rut  <br />
**nomtot**: Nombre del dueño del vehículo  <br />
**veh_rut**:  <br />
**tasac2020**: Tasación 2020  <br />

df_parque['llave_marca_modelo'] = df_parque['marca']+'-'+df_parque['modelo']

# Filtro de campos relevantes

df_parque = df_parque[['ppu','dv_ppu','tipo','ano_fab','marca','modelo','llave_marca_modelo','color','motor','chasis','rut','dv','nomtot']]

df_parque.head(5)

### Corrección Tipo Vehículos (marca modelo)

# Corrección Tipo Vehículo v2
# Agregamos Camión Liviano e identificamos nuevos minibus

df_trad2 = pd.read_excel(URL_TRAD+'Base_Traduccion_v2.xlsx',sheet_name='Hoja5')
df_trad2.head()

df_trad2 = df_trad2[['marcamodelo','tipo_veh_trabajo_2']]

df_trad2.columns = ['llave_marca_modelo','tipo_veh_trabajo_2']

# Left Join

df_ptp = pd.merge(df_parque, df_trad2, on = 'llave_marca_modelo', how = 'left')
df_ptp.head(5)

# Corrección Tipo

df_ptp.loc[df_ptp['tipo_veh_trabajo_2'].notnull(),'Tipo'] = df_ptp['tipo_veh_trabajo_2']
df_ptp.loc[df_ptp['Tipo'].isnull(),'Tipo'] = df_ptp['tipo']

#Ejemplo

df_ptp.loc[26]

# Eliminación columnas sobrantes

df_ptp = df_ptp.drop('tipo',1)
df_ptp = df_ptp.drop('tipo_veh_trabajo_2',1)

df_ptp.head()

df_ptp['ppu'].count()

### Traductor Pesados

# Traductor Pesado
# (Viene con valores duplicados)

df_tradpes = pd.read_csv(URL_TRAD+'TraduccionPesados.csv', sep = ';')
df_tradpes.columns = ['ppu','tp_llave','tp_marca_fantasia','tp_modelo_fantasia','tp_motor_fantasia','tp_marcaMotor',
                      'tp_vigencia','tipo_veh_trabajo','tp_sucursales_cobertura','tp_Marca & Modelo']

df_tradpes.head()

# Elimación de campos no relevantes

df_tradpes = df_tradpes[['ppu','tp_marca_fantasia','tp_modelo_fantasia','tp_motor_fantasia','tp_marcaMotor','tp_vigencia',
                        'tipo_veh_trabajo']]

df_tradpes.head()

# Parque con Traducción Pesados

df_ptp = pd.merge(df_ptp, df_tradpes, on = 'ppu', how = 'left')

df_ptp.head(5)

# Corrección marca

df_ptp.loc[df_ptp['tp_marca_fantasia'].isnull(),'Marca'] = df_ptp['marca']
df_ptp.loc[df_ptp['Marca'].isnull(),'Marca'] = df_ptp['tp_marca_fantasia']

df_ptp = df_ptp.drop('tp_marca_fantasia',1)
df_ptp = df_ptp.drop('marca',1)

df_ptp.head()

# Corrección tipo vehículo

df_ptp.loc[df_ptp['tipo_veh_trabajo'].isnull(),'Tipo_Final'] = df_ptp['Tipo']
df_ptp.loc[df_ptp['Tipo_Final'].isnull(),'Tipo_Final'] = df_ptp['tipo_veh_trabajo']

df_ptp = df_ptp.drop('tipo_veh_trabajo',1)
df_ptp = df_ptp.drop('Tipo',1)
df_ptp['Tipo'] = df_ptp['Tipo_Final']
df_ptp = df_ptp.drop('Tipo_Final',1)

df_ptp.head()

# Corrección modelo

df_ptp.loc[df_ptp['tp_modelo_fantasia'].isnull(),'Modelo'] = df_ptp['modelo']
df_ptp.loc[df_ptp['Modelo'].isnull(),'Modelo'] = df_ptp['tp_modelo_fantasia']

df_ptp = df_ptp.drop('tp_modelo_fantasia',1)
df_ptp = df_ptp.drop('modelo',1)

df_ptp.head()

df_ptp['Tipo'].unique()

# Actualizamos llave marca modelo

df_ptp['llave_marca_modelo'] = df_ptp['Marca']+'-'+df_ptp['Modelo']

df_ptp.head()

df_ptp['ppu'].count()

### Traducción Livianos

df_tl = pd.read_excel(URL_TRAD+'TraduccionLivianos.xlsx', sheet_name = 'Hoja1')
df_tl.columns = ['llave_livianos','tl_marca_fantasia','tl_modelo_fantasia']

df_tl.head()

df_ptp['llave_livianos'] = df_ptp['Modelo']+' '+df_ptp['Marca']

# Left Join

df_ptp = pd.merge(df_ptp, df_tl, on = 'llave_livianos', how = 'left')
df_ptp.head()

# Corrección Marca

df_ptp.loc[df_ptp['tl_marca_fantasia'].isnull(),'Marca_Final'] = df_ptp['Marca']
df_ptp.loc[df_ptp['Marca_Final'].isnull(),'Marca_Final'] = df_ptp['tl_marca_fantasia']

df_ptp['Marca'] = df_ptp['Marca_Final']
df_ptp = df_ptp.drop('Marca_Final',1)
df_ptp = df_ptp.drop('tl_marca_fantasia',1)

df_ptp.head()

# Corrección Modelo

df_ptp.loc[df_ptp['tl_modelo_fantasia'].isnull(),'Modelo_Final'] = df_ptp['Modelo']
df_ptp.loc[df_ptp['Modelo_Final'].isnull(),'Modelo_Final'] = df_ptp['tl_modelo_fantasia']

df_ptp['Modelo'] = df_ptp['Modelo_Final']
df_ptp = df_ptp.drop('tl_modelo_fantasia',1)
df_ptp = df_ptp.drop('Modelo_Final',1)

df_ptp.head()

# Eliminamos las columnas que ya no necesitamos
df_ptp = df_ptp.drop('llave_marca_modelo',1)
df_ptp = df_ptp.drop('llave_livianos',1)

df_ptp.head()

df_ptp['ppu'].count()

### Listado de Modelos a Trabajar

# Target Repuestos

df_lmat = pd.read_excel(URL_TRAD+'Listado de modelos a trabajar.xlsx', sheet_name = 'BD')
df_lmat.columns = ['Marca','Modelo','llave_marcamodelo','lmat_vigencia']

df_lmat = df_lmat[['llave_marcamodelo','lmat_vigencia']]

# Left join

df_ptp['llave_marcamodelo'] = df_ptp['Marca']+' '+df_ptp['Modelo']

df_ptp = pd.merge(df_ptp, df_lmat, on = 'llave_marcamodelo', how = 'left')

df_ptp = df_ptp.drop('llave_marcamodelo',1)

df_ptp.head()

df_ptp.to_csv('C:/Users/juan.vergara/OneDrive - Caren Repuestos Flotacentro/Documentos/Análisis/Parque Automotriz/Parque en Bruto para Modelo/ParqueAutomotriz2020.csv', index= False)
