==> Machine_Learning/Estudio Parque Automotriz/02 - ETL ConsumosTeoricos.ipynb <==
### Consumos Teóricos

import pandas as pd

# Carga Base Parque

URL = 'C:/Users/juan.vergara/OneDrive - Caren Repuestos Flotacentro/Documentos/Análisis/Parque Automotriz/Parque en Bruto para Modelo/' 

df_parque = pd.read_csv(URL + 'ParqueAutomotriz2020.csv')
df_parque.head(5)

df_parque.shape

# Consumo Teórico

df_consumo = pd.read_excel(URL + 'ConsumoTeorico.xlsx', sheet_name='Hoja2')
df_consumo.columns = ['Tipo','Consumo_Teorico_Lubricantes','Consumo_Teorico_Neumaticos','Consumo_Teorico_Repuestos']
df_consumo

# Left Join

df_parque = pd.merge(df_parque,df_consumo, on = 'Tipo', how = 'left')

df_parque.head(10)

### Tipo de Vechículo (TEMP)

df_tipo = pd.read_excel(URL + 'Dim_TipoVehiculo.xlsx', sheet_name = 'Hoja1')

df_tipo.columns = ['Tipo','tipo_corto','clasificacion']

# Left Join

df_parque = pd.merge(df_parque, df_tipo, on = 'Tipo', how = 'left')

### Corrección Consumos según Vigencias

#### Neumáticos

df_parque.loc[(df_parque['tp_vigencia'] == 'D') & (df_parque['clasificacion'] == 'PESADO'), 'Consumo_Teorico_Neumaticos'] = 0

df_parque['Consumo_Teorico_Neumaticos'] = df_parque['Consumo_Teorico_Neumaticos'].fillna(0)

df_parque['Consumo_Teorico_Neumaticos'].sum()

#### Repuestos

df_parque['Consumo_Teorico_Repuestos_TM'] = df_parque['Consumo_Teorico_Repuestos']
df_parque['Consumo_Teorico_Repuestos_TM'] = df_parque['Consumo_Teorico_Repuestos_TM'].fillna(0)

df_parque.loc[(df_parque['lmat_vigencia'] != 'V') & (df_parque['Tipo'] != 'REMOLQUE') , 'Consumo_Teorico_Repuestos'] = 0

df_parque['Consumo_Teorico_Repuestos'] = df_parque['Consumo_Teorico_Repuestos'].fillna(0)

df_parque['Consumo_Teorico_Repuestos'].sum()

df_parque['Consumo_Teorico_Repuestos_TM'].sum()

#### Lubricantes

df_parque['Consumo_Teorico_Lubricantes'] = df_parque['Consumo_Teorico_Lubricantes'].fillna(0)

df_parque['Consumo_Teorico_Lubricantes'].sum()

df_parque.dtypes

df_parque['Consumo_Teorico_Total'] = df_parque['Consumo_Teorico_Lubricantes']+df_parque['Consumo_Teorico_Neumaticos']+df_parque['Consumo_Teorico_Repuestos']

df_parque.head()

df_parque['Consumo_Teorico_Lubricantes'] = df_parque['Consumo_Teorico_Lubricantes'].astype('int64')
df_parque['Consumo_Teorico_Neumaticos'] = df_parque['Consumo_Teorico_Neumaticos'].astype('int64')
df_parque['Consumo_Teorico_Repuestos'] = df_parque['Consumo_Teorico_Repuestos'].astype('int64')
df_parque['Consumo_Teorico_Repuestos_TM'] = df_parque['Consumo_Teorico_Repuestos_TM'].astype('int64')
df_parque['Consumo_Teorico_Total'] = df_parque['Consumo_Teorico_Total'].astype('int64')

df_parque['ppu'].count()

df_parque = df_parque.drop('tipo_corto',1)
df_parque = df_parque.drop('clasificacion',1)

df_parque.to_csv('C:/Users/juan.vergara/OneDrive - Caren Repuestos Flotacentro/Documentos/Análisis/Parque Automotriz/Parque en Bruto para Modelo/ParqueAutomotriz2020_CC.csv', index= False)
