==> Machine_Learning/Estudio Parque Automotriz/03 - Rut Relacionados.ipynb <==
### Empresas Relacionadas

import pandas as pd

# Carga Base Parque

URL = 'C:/Users/juan.vergara/OneDrive - Caren Repuestos Flotacentro/Documentos/Análisis/Parque Automotriz/Parque en Bruto para Modelo/' 

df_parque2020 = pd.read_csv(URL + 'ParqueAutomotriz2020_CC.csv')

df_parque2020.head(5)

df_parque2020['ppu'].count()

# RUT relacionados

df_rutgrupos = pd.read_csv(URL + 'Rut_relacionados.csv', sep = ';')

df_rutgrupos.columns = ['rut', 'grupo','Rut_Grupo']
df_rutgrupos.head(10)

# LEFT

df_parque2020 = pd.merge(df_parque2020, df_rutgrupos, on = 'rut', how = 'left')

df_parque2020.head(5)

df_parque2020.loc[df_parque2020['Rut_Grupo'].notnull(), 'RUT_Final'] = df_parque2020['Rut_Grupo']
df_parque2020.loc[df_parque2020['RUT_Final'].isnull(),'RUT_Final'] = df_parque2020['rut']

df_parque2020[df_parque2020['RUT_Final'] == 78057000]

df_parque2020[df_parque2020['RUT_Final'] == 77542170]

# Eliminación columnas

df_parque2020 = df_parque2020.drop('Rut_Grupo',1)
df_parque2020 = df_parque2020.drop('grupo',1)

df_parque2020.head(5)

df_parque2020.dtypes

df_parque2020['RUT_Final'] = df_parque2020['RUT_Final'].astype('int64')

df_parque2020['ppu'].count()

df_parque2020.to_csv('C:/Users/juan.vergara/OneDrive - Caren Repuestos Flotacentro/Documentos/Análisis/Parque Automotriz/Parque en Bruto para Modelo/ParqueAutomotriz2020_CCR.csv', index= False)
