==> Machine_Learning/Estudio Parque Automotriz/04 - ETL Leasing y Contrato.ipynb <==
### Leasing y Contratos de Mantención

import pandas as pd

URL = 'C:/Users/juan.vergara/OneDrive - Caren Repuestos Flotacentro/Documentos/Análisis/Parque Automotriz/Parque en Bruto para Modelo/' 

df_parque2020 = pd.read_csv(URL + 'ParqueAutomotriz2020_CCR.csv')

df_parque2020.head()

df_parque2020[df_parque2020['RUT_Final'] == 77542170]

### Levantamiento 

df_levantamiento = pd.read_csv(URL+'Consolidado_Top_y_Prospectos.csv', sep = ";")

df_levantamiento.head()

df_levantamiento.columns

df_levantamiento = df_levantamiento[[' ppu ','RUT_Final',' nomtot ',' ¿Posee contrato de mantención? ',' ¿Está bajo leasing financiero? ']]

df_levantamiento.columns = ['ppu','RUT_Final_2','nomtot_Lto','mantencion','leasing']

df_levantamiento.head()

df_parque2020 = pd.merge(df_parque2020, df_levantamiento, on = 'ppu', how = 'left')

df_parque2020['mantencion'] = df_parque2020['mantencion'].fillna('SIN INFO')
df_parque2020['leasing'] = df_parque2020['leasing'].fillna('SIN INFO')

# Reemplazando los rut y nombre de empresas

df_parque2020.loc[df_parque2020['RUT_Final_2'].notnull(), 'RUT_Final'] = df_parque2020['RUT_Final_2']
df_parque2020.loc[df_parque2020['nomtot_Lto'].notnull(), 'nomtot'] = df_parque2020['nomtot_Lto']

df_parque2020['RUT_Final'] = df_parque2020['RUT_Final'].astype('int64')
df_parque2020['nomtot'] = df_parque2020['nomtot'].str.upper()

df_parque2020 = df_parque2020.drop('RUT_Final_2', axis = 1)
df_parque2020 = df_parque2020.drop('nomtot_Lto', axis = 1)

df_parque2020.head()

df_parque2020[df_parque2020['RUT_Final'] == 77542170]

df_parque2020[df_parque2020['RUT_Final'] == 78057000]

df_parque2020.to_csv('C:/Users/juan.vergara/OneDrive - Caren Repuestos Flotacentro/Documentos/Análisis/Parque Automotriz/Parque en Bruto para Modelo/ParqueAutomotriz2020_Modelo.csv', index= False)
